Make coordinator workflow more durable

+++ backend/src/Worker/Models/AgentStatusDto.cs
@@ -4,4 +4,4 @@ public record AgentStatusDto(
     AgentStatus Status,
     string? CurrentRunId,
     int QueueDepth,
-    int RunNumber);
+    long RunNumber);
